Refactor integration test framework. (#984)
* only integration

* only integration

* remove prev code

* remove prev code

* feat lint

* feat char error

* feat char error

* feat char

* test all

* make canal integration test pass

* feat unit test and inte test

* feat lint:
g

* feat unit test

* feat canal docker compose build

* feat lint

* feat import seq

* solve intrgration test

* 1.resolve conversation 2. update readme

* resolve conversation

* feat readme

* use type switch replace inherit

* rename case file
