+++ integration/framework/avro/kafka_docker_env.go
@@ -14,20 +14,15 @@
 package avro
 
 import (
-	"context"
-	"database/sql"
 	"encoding/json"
 	"errors"
-	"github.com/pingcap/ticdc/integration/framework"
 	"io/ioutil"
 	"net/http"
-	"os/exec"
 	"path"
-	"time"
 
 	"github.com/integralist/go-findroot/find"
 	"github.com/pingcap/log"
-	"github.com/pingcap/ticdc/pkg/retry"
+	"github.com/pingcap/ticdc/integration/framework"
 	"go.uber.org/zap"
 )
 
@@ -35,13 +30,11 @@ const (
 	healthCheckURI          = "http://127.0.0.1:18083"
 	dockerComposeFilePath   = "/docker-compose-avro.yml"
 	controllerContainerName = "ticdc_controller_1"
-	upstreamDSN             = "root@tcp(127.0.0.1:4000)/"
-	downstreamDSN           = "root@tcp(127.0.0.1:5000)/"
 )
 
 // KafkaDockerEnv represents the docker-compose service defined in docker-compose-avro.yml
 type KafkaDockerEnv struct {
-	framework.DockerComposeOperator
+	framework.KafkaDockerEnv
 }
 
 // NewKafkaDockerEnv creates a new KafkaDockerEnv
@@ -91,73 +84,11 @@ func NewKafkaDockerEnv(dockerComposeFile string) *KafkaDockerEnv {
 		file = dockerComposeFile
 	}
 
-	return &KafkaDockerEnv{framework.DockerComposeOperator{
-		FileName:      file,
-		Controller:    controllerContainerName,
-		HealthChecker: healthChecker,
-	}}
-}
-
-// Reset implements Environment
-func (e *KafkaDockerEnv) Reset() {
-	e.TearDown()
-	e.Setup()
-}
-
-// RunTest implements Environment
-func (e *KafkaDockerEnv) RunTest(task framework.Task) {
-	cmdLine := "/cdc " + task.GetCDCProfile().String()
-	bytes, err := e.ExecInController(cmdLine)
-	if err != nil {
-		log.Fatal("RunTest failed: cannot setup changefeed",
-			zap.Error(err),
-			zap.ByteString("stdout", bytes),
-			zap.ByteString("stderr", err.(*exec.ExitError).Stderr))
-	}
-
-	upstream, err := sql.Open("mysql", upstreamDSN)
-	if err != nil {
-		log.Fatal("RunTest: cannot connect to upstream database", zap.Error(err))
-	}
-
-	_, err = upstream.Exec("set @@global.tidb_enable_clustered_index=0")
-	if err != nil {
-		log.Info("tidb_enable_clustered_index not supported.")
-	} else {
-		time.Sleep(2 * time.Second)
-	}
-
-	downstream, err := sql.Open("mysql", downstreamDSN)
-	if err != nil {
-		log.Fatal("RunTest: cannot connect to downstream database", zap.Error(err))
-	}
-
-	taskCtx := &framework.TaskContext{
-		Upstream:   upstream,
-		Downstream: downstream,
-		Env:        e,
-		WaitForReady: func() error {
-			return retry.Run(time.Second, 120, e.HealthChecker)
+	return &KafkaDockerEnv{KafkaDockerEnv: framework.KafkaDockerEnv{
+		DockerComposeOperator: framework.DockerComposeOperator{
+			FileName:      file,
+			Controller:    controllerContainerName,
+			HealthChecker: healthChecker,
 		},
-		Ctx: context.Background(),
-	}
-
-	err = task.Prepare(taskCtx)
-	if err != nil {
-		e.TearDown()
-		log.Fatal("RunTest: task preparation failed", zap.String("name", task.Name()), zap.Error(err))
-	}
-
-	log.Info("Start running task", zap.String("name", task.Name()))
-	err = task.Run(taskCtx)
-	if err != nil {
-		e.TearDown()
-		log.Fatal("RunTest: task failed", zap.String("name", task.Name()), zap.Error(err))
-	}
-	log.Info("Finished running task", zap.String("name", task.Name()))
-}
-
-// SetListener implements Environment. Currently unfinished, will be used to monitor Kafka output
-func (e *KafkaDockerEnv) SetListener(states interface{}, listener framework.MqListener) {
-	// TODO
+	}}
 }
