+++ integration/framework/avro/kafka_single_table.go
@@ -17,7 +17,6 @@ import (
 	"bytes"
 	"database/sql"
 	"fmt"
-	"github.com/pingcap/ticdc/integration/framework"
 	"io/ioutil"
 	"net/http"
 	"time"
@@ -25,6 +24,8 @@ import (
 	"github.com/pingcap/errors"
 	"github.com/pingcap/log"
 	"go.uber.org/zap"
+
+	"github.com/pingcap/ticdc/integration/framework"
 )
 
 // SingleTableTask provides a basic implementation for an Avro test case
@@ -55,13 +56,13 @@ func (a *SingleTableTask) Prepare(taskContext *framework.TaskContext) error {
 	}
 
 	_ = taskContext.Upstream.Close()
-	taskContext.Upstream, err = sql.Open("mysql", upstreamDSN+"testdb")
+	taskContext.Upstream, err = sql.Open("mysql", framework.UpstreamDSN+"testdb")
 	if err != nil {
 		return err
 	}
 
 	_ = taskContext.Downstream.Close()
-	taskContext.Downstream, err = sql.Open("mysql", downstreamDSN+"testdb")
+	taskContext.Downstream, err = sql.Open("mysql", framework.DownstreamDSN+"testdb")
 	if err != nil {
 		return err
 	}
