+++ integration/framework/canal/kafka_docker_env_test.go
@@ -14,11 +14,11 @@
 package canal
 
 import (
-	"github.com/pingcap/ticdc/integration/framework"
 	"os/exec"
 	"testing"
 
 	"github.com/pingcap/log"
+	"github.com/pingcap/ticdc/integration/framework"
 	"github.com/stretchr/testify/require"
 )
