+++ integration/framework/kafka_docker_env.go
@@ -1,0 +1,105 @@
+// Copyright 2020 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package framework
+
+import (
+	"context"
+	"database/sql"
+	"os/exec"
+	"time"
+
+	"github.com/pingcap/log"
+	"github.com/pingcap/ticdc/pkg/retry"
+	"go.uber.org/zap"
+)
+
+const (
+	// UpstreamDSN is upstream database dsn
+	UpstreamDSN = "root@tcp(127.0.0.1:4000)/"
+	// DownstreamDSN is downstream database dsn
+	DownstreamDSN = "root@tcp(127.0.0.1:5000)/"
+)
+
+// KafkaDockerEnv represents the docker-compose service
+type KafkaDockerEnv struct {
+	DockerComposeOperator
+}
+
+// Reset implements Environment
+func (e *KafkaDockerEnv) Reset() {
+	e.TearDown()
+	e.Setup()
+}
+
+// RunTest implements Environment
+func (e *KafkaDockerEnv) RunTest(task Task) {
+	cmdLine := "/cdc " + task.GetCDCProfile().String()
+	bytes, err := e.ExecInController(cmdLine)
+	if err != nil {
+		log.Fatal("RunTest failed: cannot setup changefeed",
+			zap.Error(err),
+			zap.ByteString("stdout", bytes),
+			zap.ByteString("stderr", err.(*exec.ExitError).Stderr))
+	}
+
+	upstream, err := sql.Open("mysql", UpstreamDSN)
+	if err != nil {
+		log.Fatal("RunTest: cannot connect to upstream database", zap.Error(err))
+	}
+
+	_, err = upstream.Exec("set @@global.tidb_enable_clustered_index=0")
+	if err != nil {
+		log.Info("tidb_enable_clustered_index not supported.")
+	} else {
+		time.Sleep(2 * time.Second)
+	}
+
+	downstream, err := sql.Open("mysql", DownstreamDSN)
+	if err != nil {
+		log.Fatal("RunTest: cannot connect to downstream database", zap.Error(err))
+	}
+
+	taskCtx := &TaskContext{
+		Upstream:   upstream,
+		Downstream: downstream,
+		Env:        e,
+		WaitForReady: func() error {
+			return retry.Run(time.Second, 120, e.HealthChecker)
+		},
+		Ctx: context.Background(),
+	}
+
+	err = task.Prepare(taskCtx)
+	if err != nil {
+		e.TearDown()
+		log.Fatal("RunTest: task preparation failed", zap.String("name", task.Name()), zap.Error(err))
+	}
+
+	log.Info("Start running task", zap.String("name", task.Name()))
+	err = task.Run(taskCtx)
+	if err != nil {
+		err1 := e.DumpStdout()
+		if err1 != nil {
+			log.Warn("Failed to dump container logs", zap.Error(err1))
+		}
+		e.TearDown()
+		log.Fatal("RunTest: task failed", zap.String("name", task.Name()), zap.Error(err))
+	}
+	log.Info("Finished running task", zap.String("name", task.Name()))
+}
+
+// SetListener implements Environment. Currently unfinished, will be used to monitor Kafka output
+func (e *KafkaDockerEnv) SetListener(states interface{}, listener MqListener) {
+	// TODO
+}
