+++ integration/framework/sql_helper.go
@@ -249,13 +249,16 @@ func (r *syncSQLRequest) Send() Awaitable {
 		defer rows.Close()
 
 		if !rows.Next() {
-			log.Warn("ReadBack:", zap.Error(err))
-			return
-		}
-		r.data, err = rowsToMap(rows)
-		if err != nil {
-			log.Warn("ReadBack", zap.Error(err))
-			return
+			// Upstream does not have the row
+			if r.requestType != sqlRequestTypeDelete {
+				log.Warn("ReadBack: no row, likely to be bug")
+			}
+		} else {
+			r.data, err = rowsToMap(rows)
+			if err != nil {
+				log.Warn("ReadBack", zap.Error(err))
+				return
+			}
 		}
 
 		atomic.StoreUint32(&r.hasReadBack, 1)
@@ -325,7 +328,7 @@ func (s *sqlRequest) upsert(ctx context.Context) error {
 }
 
 func (s *sqlRequest) delete(ctx context.Context) error {
-	db, err := sqlbuilder.New("mysql", s.helper.downstream)
+	db, err := sqlbuilder.New("mysql", s.helper.upstream)
 	if err != nil {
 		return errors.AddStack(err)
 	}
@@ -447,7 +450,7 @@ func getUniqueIndexColumn(ctx context.Context, db sqlbuilder.Database, dbName st
 	row, err := db.QueryRowContext(ctx, `
 		SELECT GROUP_CONCAT(COLUMN_NAME SEPARATOR ' ') FROM INFORMATION_SCHEMA.STATISTICS
 		WHERE TABLE_SCHEMA = ? AND TABLE_NAME = ?
-		GROUP BY INDEX_NAME 
+		GROUP BY INDEX_NAME
 		ORDER BY FIELD(INDEX_NAME,'PRIMARY') DESC
 	`, dbName, tableName)
 	if err != nil {
